Added registry sub module and refactored directory hierarchy (#104)
