+++ .gitmodules
@@ -1,0 +1,9 @@
+[submodule "evernode-registry-hook"]
+	path = evernode-registry-hook
+	url = https://github.com/EvernodeXRPL/evernode-registry-hook
+[submodule "evernode-governor-hook"]
+	path = evernode-governor-hook
+	url = https://github.com/EvernodeXRPL/evernode-governor-hook.git
+[submodule "evernode-heartbeat-hook"]
+	path = evernode-heartbeat-hook
+	url = https://github.com/EvernodeXRPL/evernode-heartbeat-hook.git
