--- hook-emulator/.eslintrc.json
@@ -1,17 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "es2021": true
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "ecmaVersion": 13
-    },
-    "rules": {
-        "no-async-promise-executor": "off"
-    },
-    "globals": {
-        "__dirname": true
-    }
-}
