--- hook-emulator/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "hook-emulator",
-  "scripts": {
-    "lint": "eslint ./hook-emulator.js",
-    "build:wrapper": "gcc -m32 ./c-wrappers/main.c ../src/evernode.c ../lib/emulatorapi.c -o ./dist/hook-wrapper",
-    "build:accsetup": "npm run --prefix ../utils/account-setup build && cp -r ../utils/account-setup/dist dist/account-setup",
-    "build:self": "npm run lint && ncc build hook-emulator.js --minify -o dist",
-    "build": "npm run build:self && npm run build:wrapper && npm run build:accsetup",
-    "bundle": "rm -rf dist && npm run build && cp -r dist emulator && cp -r sec emulator/ && cp run-emulator.sh emulator/ && tar cfz dist/emulator.tar.gz emulator && rm -r emulator"
-  },
-  "dependencies": {
-    "evernode-js-client": "0.4.35",
-    "jsonwebtoken": "8.5.1",
-    "ripple-address-codec": "4.2.0",
-    "ripple-binary-codec": "1.4.0",
-    "sqlite3": "5.0.2"
-  },
-  "devDependencies": {
-    "eslint": "8.3.0"
-  }
-}
