+++ registry-index-service/package.json
@@ -7,7 +7,7 @@
         "test": "echo \"Error: no test specified\" && exit 1",
         "lint": "./node_modules/.bin/eslint ./index-manager.js",
         "build:accsetup": "npm --prefix ../utils/account-setup i && npm run --prefix ../utils/account-setup build && mkdir -p dist/account-setup && cp -r ../utils/account-setup/dist/* dist/account-setup/",
-        "build:hook": "npm --prefix ../ i && npm run --prefix ../ build && mkdir -p dist/hook-setup && cp -r ../dist/* dist/hook-setup/",
+        "build:hook": "npm --prefix ../ i && npm run --prefix ../ build && cp -r ../dist/* dist/",
         "build:self": "npm i && npm run lint && ncc build index-manager.js --minify -o dist",
         "build": "npm run build:self && npm run build:accsetup && npm run build:hook",
         "bundle": "rm -rf dist && npm run build && cp -r dist registry-index && cp index-manager.json registry-index/ && cp -r service-acc registry-index/ && cp run-registry-index.sh registry-index/ && tar cfz dist/registry-index.tar.gz registry-index && rm -r registry-index"
@@ -16,7 +16,7 @@
     "author": "",
     "license": "ISC",
     "dependencies": {
-        "evernode-js-client": "0.5.9",
+        "evernode-js-client": "0.5.11",
         "jsonwebtoken": "8.5.1",
         "fs": "0.0.1-security"
     },
