+++ registry-index-service/run-registry-index.sh
@@ -2,7 +2,9 @@
 hook_data_dir="data"
 config="accounts.json"
 account_setup="account-setup"
+grant_setup="grant-setup"
 hook_setup="hook-setup"
+hook2_setup="hook2-setup"
 index="index.js"
 data_dir=$(pwd)
 
@@ -10,11 +12,14 @@ data_dir=$(pwd)
 # Setup the dev repo paths.
 if [ ! -f $index ]; then
     account_setup="dist/account-setup"
+    grant_setup="dist/grant-setup"
     hook_setup="dist/hook-setup"
+    hook2_setup="dist/hook2-setup"
     index="index-manager.js"
 fi
 
 wasm_path="$hook_setup"
+wasm2_path="$hook2_setup"
 
 arg1=$1
 arg2=$2
@@ -60,7 +65,9 @@ function create_service() {
 }
 
 function sethook() {
+    ! CONFIG_PATH=$hook_data_dir/$arg1/$config WASM_PATH=$wasm2_path $(which node) $hook2_setup && echo "Hook2 setup faild." && exit 1
     ! CONFIG_PATH=$hook_data_dir/$arg1/$config WASM_PATH=$wasm_path $(which node) $hook_setup && echo "Hook setup faild." && exit 1
+    ! CONFIG_PATH=$hook_data_dir/$arg1/$config $(which node) $grant_setup && echo "Grant setup faild." && exit 1
 }
 
 if [ -z "$arg1" ]; then # If 1st param is empty, Create a new instance.
