+++ utils/account-setup/index.js
@@ -29,7 +29,7 @@ const ACCOUNT_DATA_DIR = process.env.ACCOUNT_DATA_DIR || __dirname;
 const HOOK_DATA_DIR = ACCOUNT_DATA_DIR + '/data'
 
 // Account names
-const accounts = ["ISSUER", "FOUNDATION_COLD_WALLET", "REGISTRY"];
+const accounts = ["ISSUER", "FOUNDATION_COLD_WALLET", "GOVERNOR", "REGISTRY", "HEARTBEAT_HOOK"];
 
 // XRP Pre-defined Special Address -> Blackhole
 const BLACKHOLE_ADDRESS = "rrrrrrrrrrrrrrrrrrrn5RM1rHd";
@@ -105,10 +105,22 @@ async function main() {
             await newAccounts[1].xrplAcc.setTrustLine(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address, TOTAL_MINTED_EVRS);
         }
 
-        const registry_lines = await newAccounts[2].xrplAcc.getTrustLines(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address);
+        const governor_lines = await newAccounts[2].xrplAcc.getTrustLines(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address);
+
+        if (governor_lines.length === 0) {
+            await newAccounts[2].xrplAcc.setTrustLine(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address, TOTAL_MINTED_EVRS);
+        }
+
+        const registry_lines = await newAccounts[3].xrplAcc.getTrustLines(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address);
 
         if (registry_lines.length === 0) {
-            await newAccounts[2].xrplAcc.setTrustLine(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address, TOTAL_MINTED_EVRS);
+            await newAccounts[3].xrplAcc.setTrustLine(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address, TOTAL_MINTED_EVRS);
+        }
+
+        const heartbeat_hook_lines = await newAccounts[4].xrplAcc.getTrustLines(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address);
+
+        if (heartbeat_hook_lines.length === 0) {
+            await newAccounts[4].xrplAcc.setTrustLine(EvernodeConstants.EVR, newAccounts[0].xrplAcc.address, TOTAL_MINTED_EVRS);
         }
 
         console.log("Trust Lines initiated");
@@ -119,7 +131,7 @@ async function main() {
 
         console.log(`${TOTAL_MINTED_EVRS} EVRs were issued to EVERNODE Foundation`);
 
-        await newAccounts[1].xrplAcc.makePayment(newAccounts[2].xrplAcc.address, REGISTRY_EVRS, EvernodeConstants.EVR, newAccounts[0].xrplAcc.address);
+        await newAccounts[1].xrplAcc.makePayment(newAccounts[3].xrplAcc.address, REGISTRY_EVRS, EvernodeConstants.EVR, newAccounts[0].xrplAcc.address);
 
         console.log(`${REGISTRY_EVRS} EVRs were transferred to Registry by the Foundation`);
         // END - Transfer Currency
@@ -152,7 +164,7 @@ async function main() {
         // END - Log Account Details
 
         // Save the generated account data in the config.
-        const configDir = path.resolve(HOOK_DATA_DIR, config.registry.address);
+        const configDir = path.resolve(HOOK_DATA_DIR, config.governor.address);
         fs.mkdirSync(configDir, { recursive: true });
         const configPath = `${configDir}/${CONFIG_FILE}`;
         console.log(`Recording account data in ${configPath}`);
